backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
datapizza-ai
anthropic==0.39.0
sqlalchemy==2.0.23
pyodbc==5.0.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
